top_cat2 sections added though they do not yet look pretty
